integration Add transation iso levels, FilterInArray

+++ mrpostgres/builder/part/sql_limit.go
@@ -9,13 +9,13 @@ import (
 type (
 	// SQLLimitBuilder - объект для создания части SQL используемой в OFFSET, LIMIT.
 	SQLLimitBuilder struct {
-		maxSize uint64
+		maxSize int
 		helper  *helper.SQLLimit
 	}
 )
 
 // NewSQLLimitBuilder - создаёт объект SQLLimitBuilder.
-func NewSQLLimitBuilder(maxSize uint64) *SQLLimitBuilder {
+func NewSQLLimitBuilder(maxSize int) *SQLLimitBuilder {
 	return &SQLLimitBuilder{
 		maxSize: maxSize,
 		helper:  helper.NewSQLLimit(),
@@ -23,7 +23,7 @@ func NewSQLLimitBuilder(maxSize uint64) *SQLLimitBuilder {
 }
 
 // Build - создаёт часть SQL, которая предназначена быть частью конкретного SQL выражения.
-func (b *SQLLimitBuilder) Build(index, size uint64) mrstorage.SQLPart {
+func (b *SQLLimitBuilder) Build(index, size int) mrstorage.SQLPart {
 	return b.createPart(b.helper.OffsetLimit(index, size, b.maxSize))
 }
